画像をgifに差し替えして、ビルド情報の記載。
