+++ Directory.Packages.props
@@ -47,6 +47,7 @@
     <PackageVersion Include="WPF-UI" Version="4.0.0" />
     <PackageVersion Include="WPF-UI.Tray" Version="3.0.5" />
     <PackageVersion Include="WpfAnalyzers" Version="4.1.1" />
+    <PackageVersion Include="XamlAnimatedGif" Version="2.3.0" />
     <PackageVersion Include="xunit" Version="2.9.3" />
     <PackageVersion Include="xunit.runner.visualstudio" Version="3.0.0" />
     <PackageVersion Include="OpenAI" Version="2.1.0" />
