+++ WindowTranslator/WindowTranslator.csproj
@@ -26,7 +26,8 @@
     <SentryUploadSources>true</SentryUploadSources>
   </PropertyGroup>
   <ItemGroup>
-    <SplashScreen Include="splash.png" />
+    <None Remove="splash.gif" />
+    <SplashScreen Include="splash.gif" />
     <Resource Include="$(ApplicationIcon)" />
     <None Include="appsettings.json" CopyToOutputDirectory="Always" />
   </ItemGroup>
@@ -58,6 +59,7 @@
     <PackageReference Include="Microsoft.VisualStudio.Threading" />
     <PackageReference Include="Weikio.PluginFramework.AspNetCore" />
     <PackageReference Include="Quickenshtein" />
+    <PackageReference Include="XamlAnimatedGif" />
   </ItemGroup>
 
   <ItemGroup>
